refactor(RdfUtils): move trusty uri checker to TrustyUriUtils class
