+++ tests/test_sign_utils.py
@@ -6,4 +6,4 @@ class TestVerifyTrusty:
 
     def test_real_uri_returns_true(self):
         np = Nanopub(source_uri="https://w3id.org/np/RAAwn-ONPeKgxxWJNBDtVotYfZcFyppCy_z8ASy2mJLKY")
-        assert verify_trusty(np.rdf, np.source_uri, np.metadata.namespace)
+        assert verify_trusty(np.rdf, np.source_uri, np.namespace)
